+++ Changelog.md
@@ -8,6 +8,7 @@
 ### Changed
 
 -   Open VSCode Button now opens VSCode in the SDK Path
+-   Warning for NCS 2.7 experimental usage with VSCode
 
 ### Fixed